+++ src/main/java/co/nvqa/common/core/cucumber/glue/DbCoreSteps.java
@@ -3,6 +3,7 @@ package co.nvqa.common.core.cucumber.glue;
 import co.nvqa.common.core.cucumber.CoreStandardSteps;
 import co.nvqa.common.core.hibernate.CodCollectionDao;
 import co.nvqa.common.core.hibernate.CodInboundsDao;
+import co.nvqa.common.core.hibernate.InboundScansDao;
 import co.nvqa.common.core.hibernate.OrderDao;
 import co.nvqa.common.core.hibernate.OrderDetailsDao;
 import co.nvqa.common.core.hibernate.OrderJaroScoresV2Dao;
@@ -10,7 +11,6 @@ import co.nvqa.common.core.hibernate.OrderTagsDao;
 import co.nvqa.common.core.hibernate.OrderTagsSearchDao;
 import co.nvqa.common.core.hibernate.OutboundScansDao;
 import co.nvqa.common.core.hibernate.ReservationsDao;
-import co.nvqa.common.core.hibernate.RouteLogsDao;
 import co.nvqa.common.core.hibernate.RouteMonitoringDataDao;
 import co.nvqa.common.core.hibernate.RouteWaypointDao;
 import co.nvqa.common.core.hibernate.ShipperPickupSearchDao;
@@ -21,7 +21,7 @@ import co.nvqa.common.core.model.order.Order;
 import co.nvqa.common.core.model.order.Order.Transaction;
 import co.nvqa.common.core.model.persisted_class.core.CodCollections;
 import co.nvqa.common.core.model.persisted_class.core.CodInbounds;
-import co.nvqa.common.core.model.persisted_class.core.CoreRouteLogs;
+import co.nvqa.common.core.model.persisted_class.core.InboundScans;
 import co.nvqa.common.core.model.persisted_class.core.OrderDetails;
 import co.nvqa.common.core.model.persisted_class.core.OrderJaroScoresV2;
 import co.nvqa.common.core.model.persisted_class.core.OrderTags;
@@ -58,8 +58,6 @@ public class DbCoreSteps extends CoreStandardSteps {
   @Inject
   private OrderDetailsDao orderDetailsDao;
   @Inject
-  private RouteLogsDao routeLogsDao;
-  @Inject
   private WaypointsDao waypointsDao;
   @Inject
   private TransactionsDao transactionsDao;
@@ -85,6 +83,8 @@ public class DbCoreSteps extends CoreStandardSteps {
   private RouteWaypointDao routeWaypointDao;
   @Inject
   private CodCollectionDao codCollectionDao;
+  @Inject
+  private InboundScansDao inboundScansDao;
 
   @Inject
   private OrderDao orderDao;
@@ -171,21 +171,6 @@ public class DbCoreSteps extends CoreStandardSteps {
     String lon = longitude.toString();
     String formattedLongitude = lon.substring(0, 6);
     return new String[]{formattedLatitude, formattedLongitude};
-  }
-
-  @When("DB Core - verify route_logs record:")
-  public void verifyRouteLogs(Map<String, String> data) {
-    Map<String, String> resolvedData = resolveKeyValues(data);
-    CoreRouteLogs expected = new CoreRouteLogs(resolvedData);
-
-    doWithRetry(() -> {
-      CoreRouteLogs actual = routeLogsDao.getRouteLogs(expected.getId());
-      Assertions.assertThat(actual)
-          .withFailMessage("Roure logs was not found: " + resolvedData);
-      Assertions.assertThat(actual).withFailMessage("Roure logs was not found: " + resolvedData)
-          .isNotNull();
-      expected.compareWithActual(actual, resolvedData);
-    }, f("verify route_logs records"), 10_000, 3);
   }
 
   @When("DB Core - verify waypoints record:")
@@ -603,6 +588,19 @@ public class DbCoreSteps extends CoreStandardSteps {
           .as("List of cod_collections records for waypointId=%s", expected.getWaypointId())
           .anyMatch(expected::matchedTo);
     }, "verify cod_collections");
+  }
+
+  @And("DB Core - Operator verifies inbound_scans record:")
+  public void verifyInboundScans(Map<String, String> data) {
+    InboundScans expected = new InboundScans(resolveKeyValues(data));
+    doWithRetry(() -> {
+      var actual = inboundScansDao.findInboundScansByOrderId(expected.getOrderId());
+      Assertions.assertThat(actual)
+          .as("List of inbound_scans records for orderId=%s", expected.getOrderId())
+          .isNotEmpty()
+          .anyMatch(expected::matchedTo);
+      putInList(KEY_CORE_LIST_OF_INBOUND_SCANS, actual);
+    }, "Fetch InboundScans", 2000, 3);
   }
 
 }